feat: addition of method Marshal for Kzg proofs
